requirements.txt
selenium
webdriver-manager
python-telegram-bot==21.9
openai
python-dotenv
Pillow